Move client to a separate file
