+++ couchdb.go
@@ -7,89 +7,8 @@ package couchdb
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
-	"net/http"
-	"net/url"
 	"strings"
 )
-
-// Client represents a remote CouchDB server.
-type Client struct{ *transport }
-
-// NewClient creates a new Client
-// addr should contain scheme and host, and optionally port and path. All other attributes will be ignored
-// If client is nil, default http.Client will be used
-// If auth is nil, no auth will be set
-func NewClient(addr *url.URL, client *http.Client, auth Auth) *Client {
-	prefixAddr := *addr
-	// cleanup our address
-	prefixAddr.User, prefixAddr.RawQuery, prefixAddr.Fragment = nil, "", ""
-	return &Client{newTransport(prefixAddr.String(), client, auth)}
-}
-
-// URL returns the URL prefix of the server.
-// The url will not contain a trailing '/'.
-func (c *Client) URL() string {
-	return c.prefix
-}
-
-// Ping can be used to check whether a server is alive.
-// It sends an HTTP HEAD request to the server's URL.
-func (c *Client) Ping() error {
-	_, err := c.closedRequest("HEAD", "/", nil)
-	return err
-}
-
-// SetAuth sets the authentication mechanism used by the client.
-// Use SetAuth(nil) to unset any mechanism that might be in use.
-// In order to verify the credentials against the server, issue any request
-// after the call the SetAuth.
-func (c *Client) SetAuth(a Auth) {
-	c.transport.setAuth(a)
-}
-
-// CreateDB creates a new database.
-// The request will fail with status "412 Precondition Failed" if the database
-// already exists. A valid DB object is returned in all cases, even if the
-// request fails.
-func (c *Client) CreateDB(name string) (*DB, error) {
-	if _, err := c.closedRequest("PUT", path(name), nil); err != nil {
-		return c.DB(name), err
-	}
-	return c.DB(name), nil
-}
-
-// CreateDBWithShards creates a new database with the specified number of shards
-func (c *Client) CreateDBWithShards(name string, shards int) (*DB, error) {
-	_, err := c.closedRequest("PUT", fmt.Sprintf("%s?q=%d", path(name), shards), nil)
-
-	return c.DB(name), err
-}
-
-// EnsureDB ensures that a database with the given name exists.
-func (c *Client) EnsureDB(name string) (*DB, error) {
-	db, err := c.CreateDB(name)
-	if err != nil && !ErrorStatus(err, http.StatusPreconditionFailed) {
-		return nil, err
-	}
-	return db, nil
-}
-
-// DeleteDB deletes an existing database.
-func (c *Client) DeleteDB(name string) error {
-	_, err := c.closedRequest("DELETE", path(name), nil)
-	return err
-}
-
-// AllDBs returns the names of all existing databases.
-func (c *Client) AllDBs() (names []string, err error) {
-	resp, err := c.request("GET", "/_all_dbs", nil)
-	if err != nil {
-		return names, err
-	}
-	err = readBody(resp, &names)
-	return names, err
-}
 
 // DB represents a remote CouchDB database.
 type DB struct {
